Updated the contact add function
The update is that the SDK now supports adding multiple groups to a
newly created contact.

The SDK doesn't work at all with the API endpoint as it stands now, this
change fixes that by forcing the user to send in an array of integers
(the API uses an array of groups) or leave it unset or set it to null.

example:

$client->contact->add('46xxxxxxxxx', 'Jane', 'Doe', array( 123, 456, 789 ));
or
$client->contact->add('46xxxxxxxxx', 'Jane', 'Doe');